UTY-699: Implement Component Translations Breakup (#115)
#### Description
1. Removed the old `ComponentTranslation` class
2. Added `ComponentDispatcherHandler`, `ComponentReplicationHandler`, and `ComponentCleanupHandler` bases classes.
3. Integrated these new systems into `SpatialOSReceiveSystem`, `SpatialOSSendSystem`,  and `CleanReactiveComponentsSystem` respectively.
4. Deleted the `WorldCommandsTranslation` class
5. Re-added the World Commands data types in `Core/Commands/WorldCommands.cs` in the namespaceing of `Improbable.Gdk.Core.Commands.CreateEntity.Request`. 
6. Driveby fixes to the loggers and the connection utils.

Also added a static ComponentId onto the `ISpatialComponentData` interface - this can easily be replaced with an attribute if needed.

**Commits are ordered logically! Review in commit order for easier reviewing!**
#### Tests
The Core is closer to compiling.. ? That's a test right? 

#### Documentation
N/A
#### Primary reviewers
If your change will take a long time to review, you can name at most two primary reviewers who are ultimately responsible for reviewing this request. @ mention them.
@zeroZshadow