Force Position to Replicate Last (#543)
#### Description
This PR forces the position component to be replicated last. Otherwise, the position change could trigger a load balancing operation that would cause valid update ops to be dropped.
#### Tests
Verified that position was hit last in the loop.
#### Documentation
N/A - Should this be documented? 
#### Primary reviewers
If your change will take a long time to review, you can name at most two primary reviewers who are ultimately responsible for reviewing this request. @ mention them.