Adding initial setup include file
the new file `initial-set-up.md` contains the setup info previously found in get-started.md.

initial-set-up.md is now the single source of truth and any new or updated setup info that applies to both the FPS starter project and the GDK in general should be put into this document. Changes made to initial-set-up will be reflected wherever it is included.