Updated upload-launch
All of the formatting.

Notes:
As part of the terminology (canonology project), we are using "start a deployment" so I have used this verb.

As part of the terminology (canonology project), we are discussing saying "deployment config file" so I described it as "deployment launch configuration file in the blurb.

The "Launch deployments" UI has lower case "d" for "deployments". This is inconsistent with the rest of the UI.

Override name, Enable simulated players <- also inconsistent.

I wasn't sure what to do with some of the settings in the Deployments window, so said to leave them as they are.